Split FileRepository out of FileService
The repository handles low level db calls,
where the service is high level logic and composition of DB & S3 actions.
