+++ src/components/file/file.module.ts
@@ -3,6 +3,7 @@ import { OrganizationModule } from '../organization';
 import { UserModule } from '../user';
 import { DirectoryResolver } from './directory.resolver';
 import { FileVersionResolver } from './file-version.resolver';
+import { FileRepository } from './file.repository';
 import { FileResolver } from './file.resolver';
 import { FileService } from './file.service';
 import { FilesBucketFactory } from './files-s3-bucket.factory';
@@ -12,6 +13,7 @@ import { FilesBucketFactory } from './files-s3-bucket.factory';
   providers: [
     DirectoryResolver,
     FilesBucketFactory,
+    FileRepository,
     FileResolver,
     FileVersionResolver,
     FileService,
